JR-4132 : Refactor names of sample projects to be more readable
