+++ JudoPayDotNetFrameworkSampleApp/Properties/AssemblyInfo.cs
